// src/app/models/user.interface.ts
export interface Users {
  iduser: number;
  numberdocument: string;
  name: string;
  lastname: string;
  address: string;
  email: string;
  numberphone: string;
  landline: string;
  datebirth: string;
  municipality: string;
  stratum: string;
  academiclevel: string;
  sex: string;
  affiliationregime: string;
  disabilitycondition: string;
  ethnicity: string;
  technologicalaccess: string;
  role: string;
  urlImage: string;
}